Remove global variables and restructure tokenizer
