+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 
 cmake_minimum_required(VERSION 3.12)
-project(bibparser)
+project(bibparser LANGUAGES CXX)
 
 
 
@@ -8,7 +8,6 @@ add_executable(bibparser
 	src/main.cpp
 	src/read_bib.cpp
 	src/tokenizer.cpp
-	src/global_variables.cpp
     src/string_utils.cpp
 )
 
@@ -16,6 +15,8 @@ target_include_directories(bibparser PRIVATE include)
 
 set_property(TARGET bibparser PROPERTY CXX_STANDARD 20)
 set(CMAKE_BUILD_TYPE Debug)
+set(CMAKE_CXX_FLAGS_INIT "-g")
+
 message("CXX Standard: ${CXX_STANDARD}")
