+++ src/main.cpp
@@ -5,7 +5,6 @@
 
 #include "read_bib.hpp"
 #include "tokenizer.hpp"
-#include "global_variables.hpp"
 #include "string_utils.hpp"
 
 
@@ -20,18 +19,9 @@ int main(int argc, char *argv[])
     tokenizer_buffer.lexeme_begin = bib_file.begin();
     tokenizer_buffer.current_char = bib_file.begin();
 
-    std::list<Token> tokens;
-    tokenizer(tokenizer_buffer, tokens);
-
-    for (Token token : tokens)
-    {
-        token.print_token();
-    }
+    tokenizer(tokenizer_buffer);
 
     std::unordered_map<std::string, TokenType> symbol_table;
-    add_identifiers(tokens, symbol_table);
-
-    
     
 }
